+++ burn-core/src/record/mod.rs
@@ -1,4 +1,5 @@
-mod state;
+mod primitive;
+mod tensor;
 
 mod base;
 mod memory;
